Allow incomplete bucket for Secure Files in object storage config
This restores the changes made in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/131819 that was
reverted in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/149479.  If
`ci_secure_files` were configured in a storage-specific settings, this
would previously disable consolidated storage settings.

This change can now go in now that
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/149480 has been
merged. That merge request relaxes the constraints when consolidated
settings can be used so that consolidated settings will still function
if `ci_secure_files` is set in storage-specific settings with the same
`connection`.

This commit applies the suggestion in
https://gitlab.com/gitlab-org/gitlab/-/issues/414673#note_1546146600
to add the `ci_secure_files` config to `SUPPORTED_TYPES` and
`ALLOWED_INCOMPLETE_TYPES` so that the consolidated storage config can
be used, but not produce an error if the bucket configuration is
missing.

This change resolves the issues that caused the original change to be
reverted https://gitlab.com/gitlab-org/gitlab/-/merge_requests/122905+
and allows us to add support for consolidated object storage without
introducing a breaking change.

Changelog: fixed