Fix flaky spec in Projects::DisableLegacyInactiveProjectsService
Most likely the spec is flaky because we assume that the first N
projects in `inactive_public_projects` are always migrated. But we don't
have any deterministic ordering in the service. Also, I have updated the
`last_activity_at` timestamp to 2 years ago as 1 year ago is just around
the cut-off time used by the service it may be the reason behind this
flaky behavior.