Move YamlProcessor::FeatureFlags to Ci::Config
We implemented YamlProcessor::FeatureFlags to be used in places where
we can't reach the context like project. We always used it in CI Entry
classes. And they are only in the Ci::Config scope. So, it is safer to
move this to Ci::Config.

The main reason we are doing this change now is that we recently had
an incident because Ci::Config was called outside of the YamlProcessor
scope and it tried to call YamlProcessor::FeatureFlags.

For the next step, we can change the class name from
YamlProcessor::FeatureFlags to Ci::Config::FeatureFlags.