Clear license caches before license reset in seat link sync
Prior to this change, when triggering seat link syncs while also
removing a license and activating an activation code in quick
succession, the Sidekiq jobs could lead to duplicated license creations
license. This is because the license cache isn't up to date when the
Sidekiq job is executed.

To resolve this issue, the seat link sync now clears the current license
cache to use the actual current license before executing the logic to
reset the license (either update the current one or create a new one)
as part of the sync.

Changelog: fixed
EE: true