Pass snapshot repository's relative path to Gitaly in tests
Gitaly passes the repository's relative path to Rails when it calls
the internal API. The relative path passed is the transaction's
snapshot repository's relative path. Without it Gitaly will not be
able to apply the repository's quarantine correctly as the quarantine
paths are relative to the snapshot repository.

While the relative path is passed in production code, tests haven't
been updated to pass the relative to Gitaly. This leads to test
failures when transactions are enabled in Gitaly.

Pass the relative path of the repository in the tests to ensure its
present when transactions are enabled. The relative path passed in
tests is the same as the actual relative path of the repository. In
reality the two paths would be different. As Gitaly is stubbed out
in the tests, it doesn't start a transaction and rewrite the relative
path so we use the original one.