Use Progressive Web App (PWA) icon as apple-touch-icon if defined
When adding GitLab to the homescreen on an Apple device the default
icon apple-touch-icon.png is used. This is hardcoded and therfore
on self-hosted instances the GitLab logo is shown.
As the pwa icon fullfills the same purpose in an standardized way,
we can use it as apple-touch-icon if defined.

Further references regarding "apple-touch-icon":
- https://developer.apple.com/library/archive/documentation/AppleApplications/Reference/SafariWebContent/ConfiguringWebApplications/ConfiguringWebApplications.html

Closes: https://gitlab.com/gitlab-org/gitlab/-/issues/358124

Changelog: added