Don't update `vulnerability_reads` records unnecessarily
The `traversal_ids` and `archived` attributes are already set for the
`vulnerability_reads` records associated with existing `vulnerability`
records so updating those records unnecessarily creates new tuples in
the heap and causes the indices to be updated associated with the old
tuples due to the MVCC design choice of PostgreSQL.
This causes an increase in the number of write-ahead logs generated.

By not updating the `vulnerability_reads` records unnecessarily, we
will reduce the amount of stress we put on the main DB as well as hot
standbys.

Changelog: performance
EE: true