Updated requesting changes blocking behaviour
With this change we create a new record against merge requests that signal that a
reviewer has requested changes.
This change means the merge request goes into a blocked state until that user
approves the merge request.
Before this change you could override this request by simply removing the user
as a reviewer.

Changelog: added