Added filtering by reviewer state to merge request finder
This is part of the new merge request dashboard we will be working on.
We need to be able to search for merge requests that have certain reviewer states.

This includes 2 different ways to query - with a reviewer and without.
With the reviewer will search for merge requests that have the assigned reviewer
with the passed in state.
Without the reviewer will search for merge requests that have any reviewer
with the passed in state. This is useful for when we will search for merge requests
a user is an assignee on and has reviews with certain states.