Improve extra_slug_path_sanitization account creation
Currently, when an account is created via oAuth (such as from BitBucket)
and the `extra_slug_path_sanitization` feature is on, there is a risk of
the sanitized username for the new user conflicting with an existing
username. In this case, the sign-up flow will fail with the
error "username is already taken."

With this change, the flow will correctly iterate the username after all
sanitization has finished, and create a new user with a de-conflicted
name. This should ensure that even similar names from other platforms
will be able to succesfully create accounts on GitLab via oAuth, SCIM,
and LDAP.

Changelog: fixed