Show membership source when the user can read group through inheritance
We are now showing the invited group members on the shared group members
page behind the `webui_members_inherited_users` feature flag. We mask
the source of membership if the current user cannot access the invited
group. We missed an edge case here where the source of membership was
being masked if the user had access to the invited group through
inheritance. Now the source of membership is not masked if the user has
access to the invited group through its parent group membership.