Increase jest timeout in CI
Since our update to node@20 we are seeing more flakiness in jest specs.
This is due to timeouts of individual specs. We currently still use the
default timeout of 5 seconds.

While we are seeing patterns of certain specs being slow because e.g.
they do big async imports, it is kind of hard to debug exactly what is
going on in CI. We have the suspicion that for example Garbage
Collection or Disk IO of specs running in parallel might be at play
here.

We are planning to create monitoring for the slowest specs and create
issues so that they can be debugged without breaking master or blocking
Merge Requests: https://gitlab.com/gitlab-org/gitlab/-/issues/457179.

This also removes the useless `slowTestThreshold` option. See
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/94416/diffs#note_1156651630.