Don't operate cross-partition in tests
Gitaly's transaction support is being rolled out. Prior to the roll
out, we want to exercise the transaction support in Rails pipeline.

Transactions introduce partitioning to Gitaly. Transactions aren't
allowed to do cross-partition operations on the file system. Object
pooling involves the file system level operations against the other
repositories in the same fork network. They need to be thus in the
same partition for the operations to work. Gitaly automatically
places the object pool and the forks in the same partition with the
origin repository as long as they are created using CreateObjectPool
and CreateFork. Rails specs are not doing this currently everywhere
which leads to test failures due to the specs attempting to do
cross-partition operations.

Update the specs to use CreateObjectPool and CreateFork where needed.