Remove unsupported GitLab versions from Gitaly docs
There is a push to remove references to unsupported GitLab versions from
the documentation. Due to the size of the undertaking, it has been split
into multiple issues. This MR resolves outdated references identified by
Vale, among others spotted manually, with changes to structure and
wording if required (for example, if the removal caused a list to have
only one bullet point).

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/443317.