Document pitfalls with upgrading PostgreSQL across glibc versions
We've had a number of customers upgrade operating systems on
PostgreSQL using replication and failover not realizing they needed
to be careful about ICU collations. This commit documents
the need to check for glibc version changes and what to do about it.

Relates to https://gitlab.com/groups/gitlab-org/-/epics/8573

Changelog: added