Resolve query performance regression on sbom reports
In [MR 147504][0] we migrated a couple of
sbom reports from fetching `vulnerability_findings` via the
`pipeline`, to using the `project` instead.

This was work related to the [epic][1] to remove the
`vulnerability_occurrence_pipelines` table.

It turns out that going through `project` is not as performant,
resulting in occasional timeout errors in SBOM ingestion.

This change re-introduces the original code path going through
`pipeline`, and puts the new code path going through the FF we are
using to roll out the epic

Resolves: https://gitlab.com/gitlab-org/gitlab/-/issues/453953

[0]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/147504
[1]:https://gitlab.com/groups/gitlab-org/-/epics/11241

Changelog: fixed
EE: true