Move relative path to HookEnv
Gitaly sends Rails the repository's relative path through
hooks. This relative path needs to be passed in quarantined
requests to Gitaly. Currently the relative path is piped
through a separate construct from the HookEnv even though
they both need to be piped to the same locations. Manage
the relative path with the HookEnv class so it is easily
settable and accessible where needed. We'll need to set
it in tests that are not currently correctly setting it.