Make PipelineScheduleType next_run_at nullable
This field represents a nullable value and requests for records
that contained null were failing with HTTP 500. This is technically
a breaking change, but with minimal impact because then number of
affected records is expected to be small. The primary source of
null values was only briefly rolled out on GitLab.com, and never on
self-managed.

See https://gitlab.com/gitlab-org/gitlab/-/issues/442235
Changelog: fixed