Fix inherited member source not visible to subgroup members
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/144638, we
introduced a bug due to which we started masking the parent group
members from the subgroup members when the visibility of these groups
were private.

To check if the member is from an ancestor we should check whether the
current source contains the member's source ID in its traversal IDs but
we were doing the opposite. The specs had a gap because we were only
testing public groups there and this problem only occurred for private
groups.

I have also added a feature spec to cover this case.

Changelog: fixed