Import epics alongside with its epic work item
When we are importing epics, either through file or direct import, we want to
create a synced work item along the epic.

This is also a preparation to keep the import functionality working as we
prepare to introduce the NOT NULL constraint on epics.issue_id column.

re https://gitlab.com/gitlab-org/gitlab/-/issues/446155