Ensure daily limit can't be zero
The daily limit is currently not validated, so it
can be any value within the underlying datatype for
the column. This can be a number larger than 1440.

When a number is greater than 1440 is configured for
the daily limit, the integer provided in the natural
expression to Fugit contains zero. This results in a
ZeroDivisionError from Fugit.

This commit ensures that the expression provided to
Fugit can't contain a zero. If the result of the
expression is less than 1, it will be treated like
a 1, eg 'every 1 minutes'.