Add support to list child pipelines in project
Currently we do not support retrieving child pipelines of a project via
the API or GraphQL. This can lead to an odd situation, that if the
parent pipeline was deleted, one cannot iterate through the child
pipelines.

We already return child pipelines if the (undocumented) `iids` parameter
is set. By following similar logic, we can just show child pipelines if
the user explicitly filters `source=parent_pipeline` which used to
always return an empty list.

Changelog: added