Avoid setting author name/email for API commits
By setting author fields, we set the authorship of a commit to
a user who didn't create this commit.

It's confusing especially when commit signing is globally enabled
for an instance. This is why we introduced a check that rejects
the commits with a custom author when commit signing is enabled.

Commit signing is globally enabled on staging and partially
enabled on production.