Show CTA to purchase more seats only to group owners
When block seat overages is enabled, group owners cannot add more
members to their group than seats in the subscription.
A CTA to purchase more seats is displayed when someone tries to add
more members to the group than seats in the subscription.
Both owners and maintainers can add members to groups.
Show the CTA to purchase more seats only to owners.