Remove `gitops_sync` usage metric counter
This change set removes the `gitops_sync` usage metric counter that was
used by the deprecated GitOps module in the GitLab Agent for Kubernetes.

In GitLab 17.0 the GitOps module in the agent will be removed and so can
this counter.

Depends on: https://gitlab.com/gitlab-org/gitlab/-/issues/406545

All associated merge requests of
https://gitlab.com/gitlab-org/gitlab/-/issues/406545 need to be merged,
released and running in production before this change set can be merged.